+++ qa-core/src/integrations/validators/postgres.spec.ts
@@ -1,10 +1,6 @@
 import { GenericContainer } from "testcontainers"
 import postgres from "../../../../packages/server/src/integrations/postgres"
 
-import { generator } from "@budibase/backend-core/tests"
-
-import couchdb from "../../../../packages/server/src/integrations/couchdb"
-
 jest.unmock("pg")
 
 describe("datasource validators", () => {
@@ -51,64 +47,6 @@ describe("datasource validators", () => {
       const result = await integration.testConnection()
       expect(result).toEqual({
         error: 'password authentication failed for user "wrong"',
-      })
-    })
-  })
-
-  describe("couchdb", () => {
-    let url: string
-
-    beforeAll(async () => {
-      const user = generator.first()
-      const password = generator.hash()
-
-      const container = await new GenericContainer("budibase/couchdb")
-        .withExposedPorts(5984)
-        .withEnv("COUCHDB_USER", user)
-        .withEnv("COUCHDB_PASSWORD", password)
-        .start()
-
-      const host = container.getContainerIpAddress()
-      const port = container.getMappedPort(5984)
-
-      await container.exec([
-        `curl`,
-        `-u`,
-        `${user}:${password}`,
-        `-X`,
-        `PUT`,
-        `localhost:5984/db`,
-      ])
-      url = `http://${user}:${password}@${host}:${port}`
-    })
-
-    it("test valid connection string", async () => {
-      const integration = new couchdb.integration({
-        url,
-        database: "db",
-      })
-      const result = await integration.testConnection()
-      expect(result).toBe(true)
-    })
-
-    it("test invalid database", async () => {
-      const integration = new couchdb.integration({
-        url,
-        database: "random_db",
-      })
-      const result = await integration.testConnection()
-      expect(result).toBe(false)
-    })
-
-    it("test invalid url", async () => {
-      const integration = new couchdb.integration({
-        url: "http://invalid:123",
-        database: "any",
-      })
-      const result = await integration.testConnection()
-      expect(result).toEqual({
-        error:
-          "request to http://invalid:123/any failed, reason: getaddrinfo ENOTFOUND invalid",
       })
     })
   })
